requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Collector & Worker (Telethon)
telethon==1.33.1

# Admin Bot (aiogram)
aiogram==3.2.0
